requirements.txt
aiohttp==3.9.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pytz==2023.3
urllib3==2.0.4
